+++ scalabel/bot/seg_dummy.py
@@ -24,7 +24,8 @@ class SegDummy(SegBase):
             x0, y0, w, h = bbox
             # create diamond inscribed in rectangle
             preds = np.array(
-                [[w / 2.0, 0], [w, h / 2.0], [w / 2.0, h], [0, h / 2.0]]
+                [[w / 2.0, 0], [w, h / 2.0], [w / 2.0, h], [0, h / 2.0]],
+                dtype=np.float64,
             )
             preds += np.array([x0, y0])
             output.append(preds.tolist())
