+++ scalabel/eval/detect.py
@@ -250,7 +250,7 @@ class COCOevalV2(COCOeval):  # type: ignore
         aind = [i for i, aRng in enumerate(p.areaRngLbl) if aRng == area_rng]
         mind = [i for i, mDet in enumerate(p.maxDets) if mDet == max_dets]
         s = self.eval[metric]
-        cat_ids = np.array(p.catIds)
+        cat_ids = np.array(p.catIds, dtype=np.int64)
         if iou_thr is not None:
             t = np.where(iou_thr == p.iouThrs)[0]
             s = s[t]
